fix: properly support selections spanning multiple elements
This is something that has annoyed me since I first wrote the
extension 3 years ago, but not enough that I've actually looked
into it until now. The problem was that the implementation
was very basic - it just wrapped your selected text in a span element,
and that was it. If your selection covered text that spanned multiple
HTML elements, then it wouldn't work. The fix I've gone with is
basically do the same "wrap-in-a-span" approach, but do it to
all the elements that you've highlighted individually.